+++ crates/proto/src/version/v800/types/biome_scatter_param.rs
@@ -1,23 +1,26 @@
 use bedrockrs_macros::ProtoCodec;
-
 use crate::v800::types::BiomeCoordinateData;
 
 #[derive(ProtoCodec, Debug, Clone)]
 pub struct BiomeScatterParamData {
-    #[endianness(le)]
+    #[vec_repr(u32)]
+    #[vec_endianness(var)]
     pub coordinates: Vec<BiomeCoordinateData>,
-    #[endianness(le)]
+    // Encoded as VarInt enum ordinal.
+    #[endianness(var)]
     pub eval_order: i32,
-    #[endianness(le)]
+    // Encoded as VarInt where -1 means "no expression op".
+    #[endianness(var)]
     pub chance_percent_type: i32,
     #[endianness(le)]
-    pub chance_percent: i32,
+    pub chance_percent: u16,
     #[endianness(le)]
     pub chance_numerator: i32,
     #[endianness(le)]
     pub change_denominator: i32,
-    #[endianness(le)]
+    // Encoded as VarInt where -1 means "no expression op".
+    #[endianness(var)]
     pub iterations_type: i32,
     #[endianness(le)]
-    pub iterations: i32,
+    pub iterations: u16,
 }
