+++ crates/proto/src/version/v800/types/biome_weighted_temperature.rs
@@ -2,8 +2,8 @@ use bedrockrs_macros::ProtoCodec;
 
 #[derive(ProtoCodec, Debug, Clone, Copy)]
 pub struct BiomeWeightedTemperatureData {
-    #[endianness(le)]
+    #[endianness(var)]
     pub temperature: i32,
     #[endianness(le)]
-    pub weight: i64,
+    pub weight: i32,
 }
